producer - Retrieve producer configurations from application.yml

+++ src/main/java/com/volkanyungul/bank_account/producer/generator/TransactionGenerator.java
@@ -1,5 +1,5 @@
 package com.volkanyungul.bank_account.producer.generator;
 
 public interface TransactionGenerator {
-    void generate(int numberOfTransactions);
+    void generate(long numberOfTransactions);
 }
